# backend/api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: yourfriends@flexiwan.com
    name: API support
    url: https://flexiwan.com
  description: This is the REST API for flexiWAN management.
  license:
    name: AGPLv3
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  termsOfService: https://flexiwan.com/legal/terms-of-service/
  title: FlexiWAN API
  version: 1.0.0
servers:
- description: Main server
  url: https://manage.flexiwan.com:443/api
paths:
...
  /devices:
    get:
      operationId: devicesGET
      parameters:
      - description: The number of items to skip before starting to collect the result
          set
        explode: true
        in: query
        name: offset
        required: false
        schema:
          type: integer
        style: form
        example: 10
      - description: Type of returned data
        explode: true
        in: query
        name: response
        required: false
        schema:
          enum:
          - detailed
          - summary
          - ids
          type: string
        example: summary
      - description: The number of items to return, total records count added to the 'records-total' header of the response
        explode: true
        in: query
        name: limit
        required: false
        schema:
          type: integer
        style: form
        example: 10
      - description: The field by which the data will be ordered
        explode: true
        in: query
        name: sortField
        required: false
        schema:
          type: string
        style: form
        example: name
      - description: Sorting order
        explode: true
        in: query
        name: sortOrder
        required: false
        schema:
          type: string
        style: form
        example: asc
      - description: Stringified array of filter objects
        explode: true
        in: query
        name: filters
        required: false
        schema:
          type: string
        style: form
        example: '[{"key":"name","op":"contains","val":"hub"}]'
      - description: Organization to be filtered by
        explode: true
        in: query
        name: org
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Device'
                type: array
                x-content-type: application/json
          description: Successful operation
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      security:
      - apiKey: []
      - bearerAuth: []
      summary: Get all registered devices
      tags:
      - Devices
      x-swagger-router-controller: Devices
      x-openapi-router-controller: DevicesController
      x-openapi-router-service: DevicesService
    delete:
      operationId: devicesDELETE
      parameters:
      - description: Organization to be filtered by
        explode: true
        in: query
        name: org
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevicesDeleteRequest'
      responses:
        "204":
          description: Successful operation
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      security:
      - bearerAuth: []
      - apiKey: []
      summary: Delete devices by IDs or matching the filter
      tags:
      - Devices
      x-swagger-router-controller: Devices
      x-openapi-router-controller: DevicesController
      x-openapi-router-service: DevicesService
  /devices/latestVersions:
    get:
      operationId: devicesLatestVersionsGET
      responses:
       "200":
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/DeviceLatestVersion'
         description: Successful operation
       "401":
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Error'
         description: Unauthorized
       "403":
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Error'
         description: Forbidden
       "404":
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Error'
         description: The specified resource was not found
       "500":
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Error'
         description: Internal server error
       default:
         content:
           application/json:
             schema:
               $ref: '#/components/schemas/Error'
         description: Unexpected error
      security:
      - apiKey: []
      - bearerAuth: []
      summary: Get devices latest available version
      tags:
      - Devices
      x-swagger-router-controller: Devices
      x-openapi-router-controller: DevicesController
      x-openapi-router-service: DevicesService
  /devices/apply:
    post:
      operationId: devicesApplyPOST
      parameters:
      - description: Organization to be filtered by
        explode: true
        in: query
        name: org
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCommand'
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyJobsResponse'
          description: Job IDs Created
          headers:
            Location:
              description: Location of a new resource created
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      security:
      - apiKey: []
      - bearerAuth: []
      summary: Execute an action on the device side
      tags:
      - Devices
      x-swagger-router-controller: Devices
      x-openapi-router-controller: DevicesController
      x-openapi-router-service: DevicesService
  /devices/upgdSched:
    post:
      operationId: devicesUpgdSchedPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevicesUpgradeRequest'
      parameters: 
      - description: Organization to be filtered by
        explode: true
        in: query
        name: org
        required: false
        schema:
          type: string
      responses:
        "201":
          content:
            application/json:
              schema:
                type: object
          description: Created
          headers:
            Location:
              description: Location of a new resource created
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "400":
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      security:
      - apiKey: []
      - bearerAuth: []
      summary: Schedule upgrade of devices
      tags:
      - Devices
      x-swagger-router-controller: Devices
      x-openapi-router-controller: DevicesController
      x-openapi-router-service: DevicesService
...
components:
...
  schemas:
...
    ApplyJobsResponse:
      properties:
        ids:
          items:
            type: integer
          type: array
        status:
          enum:
          - completed
          - partially completed
          type: string
        message:
          type: string
...
    DeviceVersion:
      example:
        agent: agent
        router: router
        device: device
      properties:
        agent:
          type: string
        device:
          type: string
        router:
          type: string
      type: object
    DeviceLatestVersion:
      example:
        "versions":
          "device": "1.1.51"
          "agent": "1.0.39"
          "vpp": "19.01-rc0"
          "frr": "7.1"
          "router": "01.01.10"
        "versionDeadline": "2020-03-15T00:00:00.000Z"
      properties:
        versions:
          $ref: '#/components/schemas/DeviceVersion'
        versionDeadLite:
          type: string
...
    DevicesUpgradeRequest:
      type: object
      properties:
        devices:
          type: array
          items:
            type: string
        date:
          type: string
...
    DeviceCommand:
      type: object
      properties:
        method:
          type: string
        devices:
          type: object
        meta:
          type: object
...
    Error:
      properties:
        code:
          type: string
        message:
          type: string
      required:
      - code
      - message
      type: object
...
    DevicesDeleteRequest:
      properties:
        ids:
          items:
            type: string
          type: array
        filters:
          items:
            $ref: '#/components/schemas/Filter'
          type: array
    Filter:
      example:
        key: name
        op: ==
        val: example
      properties:
        key:
          type: string
        op:
          type: string
        val:
          type: string
      required:
      - key
      - op
      - val
      type: object
  securitySchemes:
    apiKey:
      in: header
      name: APIKEY
      type: apiKey
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http